[feature] decrease friend’s health and happiness gradually over time

+++ src/Types.hs
@@ -23,8 +23,8 @@ module Types (WorldData(..), Direction(..), GameItem(..), Position, PositionInTi
         , playerLastDirection :: Direction
         , worldItems :: [GameItem]
         , inventory :: Int 
-        , friendHealthPercent :: Int
-        , friendHappinessPercent :: Int
+        , friendHealthPercent :: Float
+        , friendHappinessPercent :: Float
         }
 
     data Direction = DirectionUp | DirectionDown | DirectionLeft | DirectionRight deriving (Eq, Show)
